Fixes Trello bug Delete session without session token in header returns 500. https://trello.com/c/5an9cP1M

+++ svc/AdminStore/Controllers/SessionsController.cs
@@ -153,12 +153,18 @@ namespace AdminStore.Controllers
                     http.BaseAddress = new Uri(WebApiConfig.AccessControl);
                     http.DefaultRequestHeaders.Accept.Clear();
                     http.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
+                    if (Request.Headers.Contains("Session-Token") == false)
+                        throw new ArgumentNullException();
                     http.DefaultRequestHeaders.Add("Session-Token", Request.Headers.GetValues("Session-Token").First());
                     var result = await http.DeleteAsync("sessions");
                     if (result.IsSuccessStatusCode)
                         return Ok();
                     return ResponseMessage(result);
                 }
+            }
+            catch (ArgumentNullException)
+            {
+                return BadRequest();
             }
             catch
             {
